requirements.txt
python-socketio[client]==5.0.4
python-engineio==3.14.2
pymitter